+++ src/SpeakEasy.Specifications/HeadRequestSpecification.cs
@@ -1,6 +1,7 @@
 using System.Net.Http;
 using Machine.Fakes;
 using Machine.Specifications;
+using SpeakEasy.Requests;
 
 namespace SpeakEasy.Specifications
 {
