+++ src/SpeakEasy/UserAgent.cs
@@ -7,10 +7,6 @@ namespace SpeakEasy
         public static UserAgent SpeakEasy => new UserAgent(
             $"SpeakEasy/{typeof(UserAgent).GetTypeInfo().Assembly.GetName().Version}");
 
-        public UserAgent()
-        {
-        }
-
         public UserAgent(string name)
         {
             Name = name;
